feat: update extensions backup logic

+++ src/data-provider/extension-provider.ts
@@ -9,6 +9,7 @@ interface Extension extends vscode.Extension<unknown> {
     isBuiltin: boolean;
     publisher: string;
     name: string;
+    version: string;
   };
 }
 
@@ -16,25 +17,31 @@ export const ExtensionProviderId = 'extensions';
 
 export class ExtensionProvider implements DataProvider {
   readonly id = ExtensionProviderId;
-  public async backup({ path }: DataOptions): Promise<void> {
-    let installedExtensions = this.getInstalledExtensions();
-    logger.info('extensions backup', installedExtensions);
+  #extensionFolder: vscode.Uri;
 
-    if (!this.shouldRemoveExtensions) {
-      const extensions: Array<string> = await readJsonContent(this.getFilepath(path), []);
-      installedExtensions = [...extensions, ...installedExtensions].filter((v, index, arr) => arr.indexOf(v) === index);
-    }
-
-    const filename = this.getFilepath(path);
-    await writeJsonContent(filename, installedExtensions.sort());
+  constructor(extensionFolder: vscode.Uri) {
+    this.#extensionFolder = extensionFolder;
   }
 
-  private getInstalledExtensions() {
+  public async backup({ path, dryRun }: DataOptions): Promise<void> {
+    const installedExtensions = this.filterIgnoredExtensions(await this.getInstalledExtensions()).sort();
+
+    logger.info('extensions backup', installedExtensions);
+    if (!dryRun) {
+      const filename = this.getFilepath(path);
+      await writeJsonContent(filename, installedExtensions);
+    }
+  }
+
+  private async getInstalledExtensions(): Promise<Array<string>> {
+    const extensions = await readJsonContent<Array<VSCodeExtensionsJSON>>(this.#extensionFolder);
+
+    if (extensions?.length) {
+      return extensions.map(ext => ext.identifier.id);
+    }
     return vscode.extensions.all
       .filter((ext: Extension) => !ext.packageJSON.isBuiltin)
-      .map((ext: Extension) => {
-        return `${ext.packageJSON.publisher}.${ext.packageJSON.name}`;
-      });
+      .map((ext: Extension) => `${ext.packageJSON.publisher}.${ext.packageJSON.name}`);
   }
 
   private getFilepath(path: vscode.Uri) {
@@ -45,18 +52,24 @@ export class ExtensionProvider implements DataProvider {
     return !!getConfigSetting().get<boolean>('removeExtensions');
   }
 
-  public async restore({ path }: DataOptions): Promise<void> {
-    const extensions: Array<string> = await readJsonContent(this.getFilepath(path), []);
-    const installedExtensions = this.getInstalledExtensions();
+  private filterIgnoredExtensions(extensions: Array<string> | undefined): Array<string> {
+    const ignoredExtensios = getConfigSetting().get<Array<string>>('ignoreExtensions') || [];
+    return (extensions || []).filter(ext => !ignoredExtensios.includes(ext));
+  }
+
+  public async restore({ path, dryRun }: DataOptions): Promise<void> {
+    const extensions = this.filterIgnoredExtensions(await readJsonContent<Array<string>>(this.getFilepath(path)));
+    const installedExtensions = this.filterIgnoredExtensions(await this.getInstalledExtensions());
 
     const missingExtensions = extensions.filter(ext => !installedExtensions.includes(ext));
     const deletedExtensions = this.shouldRemoveExtensions
       ? installedExtensions.filter(ext => !extensions.includes(ext))
       : [];
     logger.info('extensions restore', missingExtensions, ' deleted:', deletedExtensions);
-
-    await Promise.all(missingExtensions.map(ext => this.installExtension(ext)));
-    await Promise.all(deletedExtensions.map(ext => this.deleteExtension(ext)));
+    if (!dryRun) {
+      await Promise.all(missingExtensions.map(ext => this.installExtension(ext)));
+      await Promise.all(deletedExtensions.map(ext => this.deleteExtension(ext)));
+    }
   }
 
   private async installExtension(ext: string) {
@@ -76,4 +89,12 @@ export class ExtensionProvider implements DataProvider {
       logger.error(`extension ${ext} failed to delete`, err);
     }
   }
+}
+
+export interface VSCodeExtensionsJSON {
+  identifier: {
+    id: string;
+    uuid: string;
+  };
+  version: string;
 }
