+++ src/data-provider/keybindings-provider.ts
@@ -4,17 +4,19 @@ import { logger } from '../initOutputChannel';
 
 export class KeybindingsProvider implements DataProvider {
   readonly id = 'keybindings';
-  public async backup({ path, userFolder }: DataOptions): Promise<void> {
+  public async backup({ path, userFolder, dryRun }: DataOptions): Promise<void> {
     const source = this.getFilepath(userFolder);
     const target = this.getFilepath(path);
 
-    try {
-      await vscode.workspace.fs.copy(source, target, {
-        overwrite: true,
-      });
-      logger.info('keybindings backup');
-    } catch (err) {
-      logger.error('keybindings backup failed', err);
+    logger.info('keybindings backup', source.fsPath);
+    if (!dryRun) {
+      try {
+        await vscode.workspace.fs.copy(source, target, {
+          overwrite: true,
+        });
+      } catch (err) {
+        logger.error('keybindings backup failed', err);
+      }
     }
   }
 
@@ -22,17 +24,19 @@ export class KeybindingsProvider implements DataProvider {
     return vscode.Uri.joinPath(path, 'keybindings.json');
   }
 
-  public async restore({ path, userFolder }: DataOptions): Promise<void> {
+  public async restore({ path, userFolder, dryRun }: DataOptions): Promise<void> {
     const source = this.getFilepath(path);
     const target = this.getFilepath(userFolder);
 
-    try {
-      await vscode.workspace.fs.copy(source, target, {
-        overwrite: true,
-      });
-      logger.info('keybindings restored');
-    } catch (err) {
-      logger.error('keybindings restore failed', err);
+    logger.info('keybindings restored', source.fsPath);
+    if (!dryRun) {
+      try {
+        await vscode.workspace.fs.copy(source, target, {
+          overwrite: true,
+        });
+      } catch (err) {
+        logger.error('keybindings restore failed', err);
+      }
     }
   }
 }
