+++ src/data-provider/snippets-provider.ts
@@ -4,17 +4,19 @@ import { logger } from '../initOutputChannel';
 
 export class SnippetsProvider implements DataProvider {
   readonly id = 'snippets';
-  public async backup({ path, userFolder }: DataOptions): Promise<void> {
+  public async backup({ path, userFolder, dryRun }: DataOptions): Promise<void> {
     const source = this.getFilepath(userFolder);
     const target = this.getFilepath(path);
 
-    try {
-      await vscode.workspace.fs.copy(source, target, {
-        overwrite: true,
-      });
-      logger.info('snippets backup');
-    } catch (err) {
-      logger.error('snippets backup failed', err);
+    logger.info('snippets backup', source.fsPath);
+    if (!dryRun) {
+      try {
+        await vscode.workspace.fs.copy(source, target, {
+          overwrite: true,
+        });
+      } catch (err) {
+        logger.error('snippets backup failed', err);
+      }
     }
   }
 
@@ -22,18 +24,19 @@ export class SnippetsProvider implements DataProvider {
     return vscode.Uri.joinPath(path, 'snippets');
   }
 
-  public async restore({ path, userFolder }: DataOptions): Promise<void> {
+  public async restore({ path, userFolder, dryRun }: DataOptions): Promise<void> {
     const source = this.getFilepath(path);
     const target = this.getFilepath(userFolder);
 
-    logger.info('snippets restore');
-    try {
-      await vscode.workspace.fs.copy(source, target, {
-        overwrite: true,
-      });
-      logger.info('snippets restored');
-    } catch (err) {
-      logger.error('snippets restore failed', err);
+    logger.info('snippets restore', source.fsPath);
+    if (!dryRun) {
+      try {
+        await vscode.workspace.fs.copy(source, target, {
+          overwrite: true,
+        });
+      } catch (err) {
+        logger.error('snippets restore failed', err);
+      }
     }
   }
 }
